doc: removed current documentation
The current documentation did not contain anything the README
didn't explain. If you are developing typesafety, you still can
use pydoc and the docstrings. The usage.rst was seriously out of
date. Maintaining the README file alone should not be a problem
and it is a more concise way to document the lib, since it is really
simple to use. May reintroduce it later if needed, but now there
is really not much point.