makefile: removed makefile and moved everything under tox
Since tox pretty much does everything a makefile would but in a
way better suited for Python apps, I removed the Makefile and
its mentions from other files. The only task that could not be
moved was the copyright update, but that already needed to be done
manually and we do get a proper warning beforehand.