validator: typing_inspect did not work properly with certain versions
In Python 3.5.3 and newer versions, the typing.Union is an instance of
typing._Union and was checked as such, while in previous versions, it
is the sub-class of UnionMeta. We need to support both.

The documentation includes a notice (in 3.6) that the typing module
is provisional and subject to breaks even between minor versions. That
is... not good.